Added pdb for glew. Made GLEW_STATIC defined globally for projects.Removed LNK4049 warnings for redefinition of export declarations due to some files not having GLEW_STATIC not defined.

+++ src/fli-opengl/mesh_factory.cpp
@@ -92,7 +92,7 @@ namespace opengl {
 	}
 
 	StaticMesh MeshFactory::CreateStaticMesh(const MeshDescriptor& desc) {
-		unsigned int numVertices = desc.Vertices.size();
+		unsigned int numVertices = (unsigned int)desc.Vertices.size();
 
 		std::vector<Buffer> vertexBuffers;
